Fix android build and code review issues
