+++ TransactionMobile.Maui/Pages/Transactions/MobileTopup/MobileTopupSelectProductPage.xaml.cs
@@ -6,50 +6,51 @@ using BusinessLogic.ViewModels.Transactions;
 
 public partial class MobileTopupSelectProductPage : ContentPage
 {
-    private MobileTopupSelectProductPageViewModel viewModel => BindingContext as MobileTopupSelectProductPageViewModel;
+    #region Constructors
 
-	public MobileTopupSelectProductPage(MobileTopupSelectProductPageViewModel vm)
-	{
-		InitializeComponent();
-        BindingContext = vm;
+    public MobileTopupSelectProductPage(MobileTopupSelectProductPageViewModel vm) {
+        this.InitializeComponent();
+        this.BindingContext = vm;
     }
 
-    protected override async void OnAppearing()
-    {
+    #endregion
+
+    #region Properties
+
+    private MobileTopupSelectProductPageViewModel viewModel => this.BindingContext as MobileTopupSelectProductPageViewModel;
+
+    #endregion
+
+    #region Methods
+
+    protected override async void OnAppearing() {
         base.OnAppearing();
-        await viewModel.Initialise(CancellationToken.None);
-        this.LoadProducts(viewModel);
-        
+        await this.viewModel.Initialise(CancellationToken.None);
+        this.LoadProducts(this.viewModel);
     }
 
-    private void LoadProducts(MobileTopupSelectProductPageViewModel viewModel)
-    {
+    private void LoadProducts(MobileTopupSelectProductPageViewModel viewModel) {
         this.ProductsList.Clear();
-        
+
         Int32 rowCount = 0;
-        foreach (ContractProductModel modelProduct in viewModel.Products)
-        {
-            Button button = new Button
-            {
-                                  Text = modelProduct.ProductDisplayText,
-                                  HorizontalOptions = LayoutOptions.FillAndExpand,
-                                  AutomationId = modelProduct.ProductDisplayText,
-            };
+        foreach (ContractProductModel modelProduct in viewModel.Products) {
+            Button button = new Button {
+                                           Text = modelProduct.ProductDisplayText,
+                                           HorizontalOptions = LayoutOptions.FillAndExpand,
+                                           AutomationId = modelProduct.ProductDisplayText,
+                                       };
             button.SetDynamicResource(VisualElement.StyleProperty, "MobileTopupButtonStyle");
 
-            Binding commandParameter = new Binding()
-                                       {
-                                           Source = new ItemSelected<ContractProductModel>
-                                                    {
-                                                        SelectedItem = modelProduct,
-                                                        SelectedItemIndex = rowCount
-                                                    }
-                                       };
+            Binding commandParameter = new Binding {
+                                                       Source = new ItemSelected<ContractProductModel> {
+                                                                                                           SelectedItem = modelProduct,
+                                                                                                           SelectedItemIndex = rowCount
+                                                                                                       }
+                                                   };
 
-            Binding command = new Binding
-                              {
-                                  Source = viewModel.ProductSelectedCommand
-                              };
+            Binding command = new Binding {
+                                              Source = viewModel.ProductSelectedCommand
+                                          };
 
             button.SetBinding(Button.CommandProperty, command);
             button.SetBinding(Button.CommandParameterProperty, commandParameter);
@@ -59,4 +60,6 @@ public partial class MobileTopupSelectProductPage : ContentPage
             rowCount++;
         }
     }
+
+    #endregion
 }
